ref: moved Customer schema into /schemas/customer
