+++ functions/invoices/getBuyHistory.ts
@@ -1,7 +1,7 @@
 import { Article } from '@/schemas/Article'
 import { getInvoices } from '@/functions/invoices/getInvoices'
 import { Invoice, InvoiceItem } from '@/schemas/Invoice'
-import { Customer } from '@/schemas/Customer'
+import { Customer } from '@/schemas/customer/Customer/Customer/Customer'
 
 /**
  * This function returns the buy-history for given array of article-ids.
